Added test_swarm_installation

+++ src/cux/cux.cpp
@@ -438,7 +438,7 @@ bool cux_init()
 
 #if !__DEVICE_EMULATION__
 	// Memory info
-	size_t free = 0, total = 0;
+	unsigned int free = 0, total = 0;
 	cuxErrCheck( (cudaError)cuMemGetInfo(&free, &total) );
 //	MLOG(verb2) << "Device memory (free, total): " << free / (1<<20) << "M, " << total / (1<<20) << "M" << "\n";
 #endif
